requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
selenium==4.15.0
# pydantic==2.5.0
gspread==5.12.0
google-auth==2.23.4
python-dotenv==1.0.0
webdriver-manager==4.0.1
apscheduler==3.10.4